Moved all functions required for unpack_file_url() from download.py to unpacker.py
